requirements.txt
anthropic>=0.8.1
python-dotenv==1.0.1
pytest==8.0.2
typing-extensions==4.9.0 